abstract out test
